+++ src/main/java/com/zontik/groshiky/service/AccountService.java
@@ -11,7 +11,7 @@ import org.springframework.transaction.annotation.Transactional;
 public class AccountService implements IAccountService{
 
     private final AccountRepository accountRepository;
-    private final static float DEFAULT_BALANCE = 0f;
+    private final static double DEFAULT_BALANCE = 0d;
 
     @Autowired
     public AccountService(AccountRepository accountRepository) {
